# public/swagger.yaml
openapi: 3.0.3
info:
  title: Vinci Barbearia API
  description: |
    API completa para gerenciamento de barbearia incluindo agendamentos, barbeiros, clientes, produtos, planos e webhooks.

    ## Autenticação
    A maioria dos endpoints requer autenticação via Bearer Token no header:
    ```
    Authorization: Bearer SEU_TOKEN_AQUI
    ```

    ## Formato de Datas
    - Datas são no formato brasileiro: DD/MM/YYYY ou YYYY-MM-DD
    - Horários são no formato: HH:MM (24h)

    ## Webhooks
    O sistema suporta dois tipos de webhooks:
    1. **Webhook Global**: Configurado na tabela `configuracoes`, dispara para todos os eventos
    2. **Webhooks de Barbeiros**: Webhooks personalizados por profissional

  version: 3.0.0
  contact:
    name: Vinci Barbearia
    url: https://vincebarbearia.vercel.app
servers:
  - url: https://vincebarbearia.vercel.app/api
    description: Produção
  - url: http://localhost:3000/api
    description: Desenvolvimento
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token de autenticação da API
  schemas:
...
    Agendamento:
      type: object
      properties:
        id:
          type: string
          format: uuid
        cliente_id:
          type: string
          format: uuid
        profissional_id:
          type: string
          format: uuid
        data_agendamento:
          type: string
          example: "19/12/2025"
        hora_inicio:
          type: string
          example: "14:30"
        hora_fim:
          type: string
          example: "15:30"
        nome_cliente:
          type: string
        telefone:
          type: string
        valor:
          type: number
          format: float
        status:
          type: string
          enum: [agendado, confirmado, em_andamento, concluido, cancelado]
        observacoes:
          type: string
...
paths:
...
  /agendamentos:
    get:
      tags:
        - Agendamentos
      summary: Listar agendamentos
      description: Retorna lista de agendamentos com filtros opcionais
      parameters:
        - name: data
          in: query
          schema:
            type: string
          description: Filtrar por data específica (formato brasileiro DD/MM/YYYY)
        - name: profissional_id
          in: query
          schema:
            type: string
            format: uuid
          description: Filtrar por profissional
      responses:
        '200':
          description: Lista de agendamentos retornada com sucesso
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Agendamento'
...
  /barbeiro/agendamentos:
    get:
      tags:
        - Barbeiros
      summary: 🆕 Agendamentos do Barbeiro (Recomendada para IA/Chatbots)
      description: |
        **⭐ API NOVA E RECOMENDADA** para barbeiros consultarem agendamentos com linguagem natural.

        Ideal para integração com IA, chatbots e WhatsApp!

        **Filtros aceitos no parâmetro `quando`:**
        - **Dias relativos:** `hoje`, `amanha`
        - **Dias da semana:** `segunda`, `terca`, `quarta`, `quinta`, `sexta`, `sabado`, `domingo`
        - **Data específica:** `21/12/2024` ou `2024-12-21`
        - **Sem filtro:** retorna todos os agendamentos futuros

        **Perguntas que a IA pode responder:**
        - "Quais meus agendamentos hoje?" → `quando=hoje`
        - "Tenho cliente amanhã?" → `quando=amanha`
        - "Quantos clientes tenho na terça?" → `quando=terca`
        - "Mostra minha agenda de quinta" → `quando=quinta`
        - "Agendamentos do dia 25/12" → `quando=25/12/2024`
        - "Quais meus próximos clientes?" → sem parâmetro `quando`

        **Exemplo de uso:**
        ```bash
        # Agendamentos de hoje
        curl -X GET 'https://vincebarbearia.vercel.app/api/barbeiro/agendamentos?barbeiro=Hiago&quando=hoje' \
          -H 'Authorization: Bearer SEU_TOKEN'

        # Próxima terça-feira
        curl -X GET 'https://vincebarbearia.vercel.app/api/barbeiro/agendamentos?barbeiro=Hiago&quando=terca' \
          -H 'Authorization: Bearer SEU_TOKEN'
        ```

        **Diferenças em relação à API antiga:**
        - ✅ Aceita dias da semana naturais (terca, quinta, etc.)
        - ✅ Mais simples: apenas `barbeiro` e `quando`
        - ✅ Sem `quando`: retorna todos futuros automaticamente
        - ✅ Aceita datas em múltiplos formatos

        📖 **Documentação completa:** Ver arquivo `API-BARBEIRO-AGENDAMENTOS.md`
      security:
        - BearerAuth: []
      parameters:
        - name: barbeiro
          in: query
          required: true
          schema:
            type: string
          description: Nome do barbeiro OU UUID do barbeiro
          example: "Hiago"
          examples:
            nome:
              value: "Hiago"
              summary: Buscar por nome
            uuid:
              value: "1039a091-b264-4c17-8fd6-88732f2112aa"
              summary: Buscar por UUID
        - name: quando
          in: query
          required: false
          schema:
            type: string
          description: |
            Filtro de data com linguagem natural:
            - Dias: `hoje`, `amanha`
            - Semana: `segunda`, `terca`, `quarta`, `quinta`, `sexta`, `sabado`, `domingo`
            - Data: `21/12/2024` ou `2024-12-21`
            - Vazio: todos os agendamentos futuros
          example: "terca"
      responses:
        '200':
          description: Agendamentos retornados com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      barbeiro:
                        type: object
                        properties:
                          id:
                            type: string
                            format: uuid
                          nome:
                            type: string
                            example: "Hiago"
                      filtro:
                        type: string
                        example: "terca"
                      descricao:
                        type: string
                        example: "terça-feira (24/12/2024)"
                      data_filtro:
                        type: string
                        example: "24/12/2024"
                      total_agendamentos:
                        type: integer
                        example: 5
                      valor_total:
                        type: number
                        format: float
                        example: 250.00
                      agendamentos:
                        type: array
                        items:
                          type: object
                          properties:
                            id:
                              type: string
                              format: uuid
                            data:
                              type: string
                              example: "24/12/2024"
                            hora:
                              type: string
                              example: "09:00"
                            cliente:
                              type: string
                              example: "João Silva"
                            telefone:
                              type: string
                              example: "11999999999"
                            servicos:
                              type: string
                              example: "Corte + Barba"
                            valor:
                              type: number
                              format: float
                              example: 50.00
                            status:
                              type: string
                              example: "confirmado"
                            observacoes:
                              type: string
                              nullable: true
                      mensagem_whatsapp:
                        type: string
                        description: Mensagem formatada para enviar no WhatsApp
        '400':
          description: Parâmetro inválido ou faltando
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: 'Parâmetro "barbeiro" é obrigatório'
        '401':
          description: Token não fornecido
        '403':
          description: Token inválido
        '404':
          description: Barbeiro não encontrado
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: 'Barbeiro "NomeInexistente" não encontrado'
        '500':
          description: Erro interno do servidor
  /barbeiros/meus-agendamentos:
    get:
      tags:
        - Barbeiros
      summary: Meus agendamentos do barbeiro (Legacy)
      description: |
        Retorna os agendamentos do barbeiro (para usar no WhatsApp).

        **⚠️ ATENÇÃO:** Para IA/Chatbots, use a nova API `/barbeiro/agendamentos` (recomendada)

        **Períodos disponíveis:**
        - hoje
        - amanha
        - semana
        - semana_que_vem
        - mes
        - mes_que_vem
        - proximos7dias
        - proximos30dias

        **Exemplo de uso:**
        ```bash
        curl -X GET 'https://vincebarbearia.vercel.app/api/barbeiros/meus-agendamentos?barbeiro_nome=Hiago&periodo=hoje' \
          -H 'Authorization: Bearer SEU_TOKEN'
        ```
      security:
        - BearerAuth: []
      parameters:
        - name: barbeiro_nome
          in: query
          required: true
          schema:
            type: string
          description: Nome do barbeiro
          example: "Hiago"
        - name: periodo
          in: query
          schema:
            type: string
            default: hoje
            enum: [hoje, amanha, semana, semana_que_vem, mes, mes_que_vem, proximos7dias, proximos30dias]
          description: Período dos agendamentos
        - name: data_inicio
          in: query
          schema:
            type: string
          description: Data de início customizada (DD-MM-YYYY ou YYYY-MM-DD)
        - name: data_fim
          in: query
          schema:
            type: string
          description: Data fim customizada (DD-MM-YYYY ou YYYY-MM-DD)
      responses:
        '200':
          description: Agendamentos retornados
...
  /clientes/meus-agendamentos:
    get:
      tags:
        - Clientes
      summary: Meus agendamentos do cliente
      description: |
        Retorna os agendamentos FUTUROS de um cliente pelo telefone.

        **⚠️ IMPORTANTE: Envie o telefone via header OU query parameter**

        **Exemplo com HEADER (recomendado para N8N):**
        ```bash
        curl -X GET 'https://vincebarbearia.vercel.app/api/clientes/meus-agendamentos' \
          -H 'Authorization: Bearer SEU_TOKEN' \
          -H 'telefone: 11970307000'
        ```

        **Exemplo com QUERY:**
        ```bash
        curl -X GET 'https://vincebarbearia.vercel.app/api/clientes/meus-agendamentos?telefone=11970307000' \
          -H 'Authorization: Bearer SEU_TOKEN'
        ```
      security:
        - BearerAuth: []
      parameters:
        - name: telefone
          in: header
          required: true
          schema:
            type: string
          description: Telefone do cliente (OBRIGATÓRIO - pode ser enviado via header ou query)
          example: "11970307000"
      responses:
        '200':
          description: Agendamentos futuros do cliente
          content:
            application/json:
              schema:
                type: object
                properties:
                  cliente:
                    type: object
                    properties:
                      telefone:
                        type: string
                      nome:
                        type: string
                  total_agendamentos:
                    type: integer
                  proximo_agendamento:
                    type: object
                  agendamentos_futuros:
                    type: array
                    items:
                      type: object
              example:
                cliente:
                  telefone: "11970307000"
                  nome: "João Silva"
                total_agendamentos: 2
                proximo_agendamento:
                  id: "uuid-123"
                  data: "20/12/2025"
                  hora_inicio: "14:30"
                  barbeiro: "Hiago"
                  pode_cancelar: true
                agendamentos_futuros:
                  - id: "uuid-123"
                    data: "20/12/2025"
                    hora_inicio: "14:30"
                    barbeiro: "Hiago"
        '400':
          description: Telefone não fornecido

  # =====================
  # SERVIÇOS
  # =====================
...